sysd: Split logic module into submodules
